linux-local: Add variable for local defconfig with absoluth path
LOCAL_DEFCONFIG variable can be used in local.conf to specify an
absolute path to a file that will be used as the kernel's defconfig.
The file can have any name, as it will be automatically renamed.

If not set, a Linux configuration file need to be copied next to
linux-local.bb recipe: meta-intel-distro/recipes-kernel/linux/defconfig.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>